# Google-Cloud-Extended-Next-2026/rewrite/openapi.yaml
openapi: 3.1.0
info:
  title: High-Concurrency Concert Ticket Booking API
  version: 1.0.0
  description: Modernized decoupled API contract matching legacy system
servers:
  - url: /api/v1
paths:
  /checkout:
    post:
      summary: Process a ticket checkout
      operationId: processCheckout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - event_id
                - ticket_type
                - quantity
                - payment_token
              properties:
                event_id:
                  type: integer
                ticket_type:
                  type: string
                quantity:
                  type: integer
                  minimum: 1
                payment_token:
                  type: string
      responses:
        '200':
          description: Order completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Order completed successfully
                  data:
                    type: object
                    properties:
                      order_id:
                        type: integer
                      total_amount:
                        type: number
                        format: float
                      status:
                        type: string
                        example: PAID
                      tickets:
                        type: object
                        properties:
                          event:
                            type: string
                          type:
                            type: string
                          quantity:
                            type: integer
        '400':
          description: Bad Request / Validation / Payment Failure
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /orders/{id}:
    get:
      summary: Get order details
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order details with user and payments
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    type: object
                  user:
                    type: object
                  payments:
                    type: array
                    items:
                      type: object
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string